Admin Backend done

+++ src/Components/Admin/teams/CreateTeam.tsx
@@ -22,9 +22,18 @@ const CreateTeam=()=> {
   }, []);
 
   const handleDelete=async(id:number)=>{
-
+ Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+if (result.isConfirmed) {
     try {
-      const response = await axios.delete(
+      const response =  axios.delete(
         `https://apexwpc.apextechno.co.uk/api/teams/${id}`,
 
         {
@@ -37,7 +46,10 @@ const CreateTeam=()=> {
      const filterPermission=permission.filter((per:any)=>per.id!==id);
      setPermission(filterPermission);
     } catch (error) {}
-  }
+   }
+  });
+  };
+
   const getPermissions = async () => {
     try {
       const response = await axios.get(
